create and use CreateHardwareBufferCommandHeader2

+++ framework/decode/api_decoder.h
@@ -85,7 +85,7 @@ class ApiDecoder
                                         uint32_t                                            width,
                                         uint32_t                                            height,
                                         uint32_t                                            stride,
-                                        uint32_t                                            usage,
+                                        uint64_t                                            usage,
                                         uint32_t                                            layers,
                                         const std::vector<format::HardwareBufferPlaneInfo>& plane_info) = 0;
