+++ framework/decode/vulkan_decoder_base.cpp
@@ -107,7 +107,7 @@ void VulkanDecoderBase::DispatchCreateHardwareBufferCommand(
     uint32_t                                            width,
     uint32_t                                            height,
     uint32_t                                            stride,
-    uint32_t                                            usage,
+    uint64_t                                            usage,
     uint32_t                                            layers,
     const std::vector<format::HardwareBufferPlaneInfo>& plane_info)
 {
