+++ framework/decode/vulkan_replay_consumer_base.cpp
@@ -420,7 +420,7 @@ void VulkanReplayConsumerBase::ProcessCreateHardwareBufferCommand(
     uint32_t                                            width,
     uint32_t                                            height,
     uint32_t                                            stride,
-    uint32_t                                            usage,
+    uint64_t                                            usage,
     uint32_t                                            layers,
     const std::vector<format::HardwareBufferPlaneInfo>& plane_info)
 {
